requirements.txt
PyYAML
paramiko
apache-libcloud==1.5.0
click